[Added] Destroy mission configurability from mission_manager.lua
Both the destroy mission distance and the reward are now configurable from the
mission_manager.lua file. Both values have four attributes to configure with
the following forumulas:
distance = base + factor * difficulty + rand(rBase) + rand(rFactor * difficulty)
reward = base + factor * difficulty + rand(rBase) + rand(rFactor * difficulty)

Settings configured to be as identical with previous values as possible.
The distance calcuation is the same but the reward is "nerfed" a bit. The
maximum payout is lowered by 150 credits. These 150 credits are randomized in
the previous implementation resulting in an average reward drop of 75 credits.

Change-Id: I399084228c9aa6be704da4ca9b0f8d8170153dcc