[Fixed] Command execution not respecting command delay and order.
This is a fix for Mantis-7793.

The issue was due to the command element being removed from the
command queue before command execution. If the command itself required
a crosslock there was a chance for another of the commands to sneak
through and get the creature object lock and start the execution of
next command as if the queue was empty.

Waiting with removing the command from the queue until after the
execution enforces any follow up commands to enqueue the command in
the queue instead of scheduling it for immediate execution.

Change-Id: I6c7edc50124622c7a08981b3fbbf6919530f8f3c