[fixed] vendor item relist issues after hitting max items - Mantis #8229
Main problem lies within the problem of ManagedWeakReference and the ==
operator behavior.  The comparing process may or may not return the
right result depending on the correctness of the savedObjectID of
ManagedWeakReference.

A case where the savedObjectID will not be correct:

An object id is 0 until ObjectManager::instance()->persistObject is
called where it first gets on the object. Constructing a
ManagedWeakReference from such an object results in the savedObjectID
to also be 0.

The fix contains a reordering of the persist call and creation of
ManagedWeakReference for the commoditiesList.

A proper fix for the general issue is also in work and will follow.

Change-Id: I8200383c7a2a1e5f152637c5077cfb6f887f7ffd