Adding 24 Faction swap to echo base, pvp zone, and more npcs (#329)
Added a 24 hour Faction swap to Echo base and initialized it as a GCW base. Added some null checking in GCW manager, created a pvp zone with a SUI that pops up on enter, with a bounce back like the village. Added a few more NPCS for testing and validation.