Loot changes/asifab (#503)
* Update rancor_tooth.lua

Increased the min and max damage on rancor teeth. Not sure if this is to high of an increase or not. Still trying to learn. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update rancor_bile.lua

Changed the charges min/max from 70 min to 90 min and max from 140 to 150. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update rancor_padded_armor_segment.lua

Changed the armor effectiveness chances to 20% min and 35% max. 

This should make it so the segments could actually be used in creating padded armor that players would use. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update rancor_hides.lua

Changed the armor effectiveness to 4% min and 8% max, from 3% min and 6% max. 

This would allow for more of them to be used in armor. As the only hides that are used are 6% or better. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update rancor_tooth.lua

The gemini bot caught that my previous change could of created an instance where the min rolled higher than the max. 

This will fix that issue. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update endor_world.lua

Adjusted Blacksun spawn weight from 3 to 5 per Desporo. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update dathomir_world.lua - increase blacksun spawn rate

Adjusted Blacksun spawn weight from 3 to 5 per Desporo. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update yavin4_world.lua - increase blacksun spawn rate

Adjusted Blacksun spawn weight from 3 to 5 per Desporo. 

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

---------

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>