Tantel armor change hoth (#531)
* Update armor_tantel_skreej_boots.lua

Added cold protection to special resists, to be ready for hoth in future.

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update armor_tantel_skreej_chest_plate.lua

Added cold protection to special resists, to be ready for hoth in future.

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update armor_tantel_skreej_helmet.lua

Added cold protection to special resists, to be ready for hoth in future.

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update armor_tantel_skreej_boots_quest.lua

Added cold to special protection for hoth

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update armor_tantel_skreej_chest_plate_quest.lua

Added cold to special protection for hoth

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

* Update armor_tantel_skreej_helmet_quest.lua

Added cold to special protection for hoth

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>

---------

Signed-off-by: Asifab <54614510+Asifab19@users.noreply.github.com>